requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
black==23.10.0
boto3==1.26.135
botocore==1.29.165
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.5
decorator==5.1.1
distlib==0.3.7
dj-database-url==2.0.0
Django==4.2.1
django-configurations==2.4.1
django-extensions==3.2.3
django-filter==23.2
django-model-utils==4.3.1
django-nose==1.4.7
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-storages==1.13.2
django-unique-upload==0.2.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
drf-yasg==1.21.7
executing==2.0.0
factory-boy==3.2.1
Faker==19.9.0
filelock==3.12.4
flake8==6.0.0
ghp-import==2.1.0
gunicorn==20.1.0
identify==2.5.30
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.13.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.5.3
mock==5.0.2
mypy-extensions==1.0.0
newrelic==8.8.0
nodeenv==1.8.0
nose==1.3.7
openapi-codec==1.3.2
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.5.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.16.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
pyyaml_env_tag==0.1
requests==2.31.0
s3transfer==0.6.2
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.17
utils==1.0.1
virtualenv==20.24.5
watchdog==3.0.0
wcwidth==0.2.8
